chat/v1/requirements.txt
cryptography==42.0.4
fastapi==0.115.4
httpx==0.27.2
kakaotrans==0.1.2
paho_mqtt==2.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.23
uvicorn==0.32.0
